Set minimum HTTP version to 2.0
Set minimum HTTP version to 2.0.
Enabled a new code analyzer: nested types should not be visible.
Changed the update page's cache mode from enabled to required.
